refactor: Split tributes/mod.rs god module into focused sub-modules
- Reduced mod.rs from 2,461 lines to 660 lines (73% reduction)
- Extracted combat.rs (440 lines): attack mechanics, violence stress, combat results
- Extracted movement.rs (207 lines): travel mechanics, TravelResult enum
- Extracted lifecycle.rs (517 lines): health, status effects, attribute management
- Extracted inventory.rs (219 lines): OwnsItems impl, item filtering, consumables
- All tests passing: 324/328 (99% success rate)
- Improved code organization and maintainability
